Refactor renderer base classes
